Improve DocumentFormat and add Pair components' handlers
This commit includes enhancements in DocumentFormat and addition of operators for Pair components. The modifications in DocumentFormat make it stricter by enforcing checks on map keys and making sure they are of type String. The commit also adds operators for first and second components of a Pair.

Update JVM vendor in buildlogic file

This commit changes the JVM vendor specification in the nitrite-java's build-logic file. Previously, the vendor was set to AdoptOpenJdk, and this

Add Foojay's Java toolchain resolver to the build configuration

To ensure we use a consistent JDK version across different build environments, we've added Foojay's Java toolchain resolver to the build configuration. Specifically, version 0.8.0 was added to settings.gradle.kts. Furthermore, toolchain settings were added to java-conventions.gradle.kts to specify language version 17, and AdoptOpenJDK as the JVM vendor.

Update Kotlin serialization and repository handling

This commit involves updates in Kotlin serialization and repository. Changes include the introduction of a DocumentFormat file, enhancement in DocumentEncoder and DocumentDecoder files, and other minor bug fixes in different modules. The update queried check on encoded strings for test mapping and precise adjustments on nitrite collection and repository handling.

Added nitrite-java submodule to fix repo issue