Add feature to disable repository type validation
A new feature has been implemented that allows developers to disable the repository type validation in the Nitrite database. This is particularly useful when dealing with objects that cannot be converted to a Nitrite Document, as it prevents Nitrite from throwing an exception. The feature is enabled by default and can be disabled using the `disableRepositoryTypeValidation()` method in the `NitriteBuilder` class. This change also includes updates to the respective test cases and documentation.