Add serialization support for NitriteId
A new serializer has been added for NitriteId, making it possible to serialize and deserialize this object type. To accommodate this change, some updates were necessary in KotlinXSerializationMapperTest where the "_id" field now receives a NitriteId instead of a string. Changes made should increase the flexibility and capabilities of the library.